Pull out Buildable into its own file/header in libnixstore
