+++ src/libcmd/installables.hh
@@ -2,12 +2,11 @@
 
 #include "util.hh"
 #include "path.hh"
+#include "buildable.hh"
 #include "eval.hh"
 #include "flake/flake.hh"
 
 #include <optional>
-
-#include <nlohmann/json_fwd.hpp>
 
 namespace nix {
 
@@ -15,25 +14,6 @@ struct DrvInfo;
 struct SourceExprCommand;
 
 namespace eval_cache { class EvalCache; class AttrCursor; }
-
-struct BuildableOpaque {
-    StorePath path;
-    nlohmann::json toJSON(ref<Store> store) const;
-};
-
-struct BuildableFromDrv {
-    StorePath drvPath;
-    std::map<std::string, std::optional<StorePath>> outputs;
-    nlohmann::json toJSON(ref<Store> store) const;
-};
-
-typedef std::variant<
-    BuildableOpaque,
-    BuildableFromDrv
-> Buildable;
-
-typedef std::vector<Buildable> Buildables;
-nlohmann::json buildablesToJSON(const Buildables & buildables, ref<Store> store);
 
 struct App
 {
